Added API authentication

+++ app/api/controllers.py
@@ -1,15 +1,43 @@
-from flask import Blueprint, jsonify, request, send_from_directory
+from functools import wraps
 
-from app import data_loader, date_time_transformer, data_transformer, numerical_transformer, one_hot_encoder, UPLOAD_FOLDER
+from flask import abort, Blueprint, jsonify, request, send_from_directory
+from flask_login import current_user, login_user
+from passlib.hash import sha256_crypt
+
+from app import data_loader, date_time_transformer, data_transformer, numerical_transformer, one_hot_encoder, \
+    UPLOAD_FOLDER
 from app.history.models import History
+from app.user_service.models import UserDataAccess
 
 api = Blueprint('api', __name__)
 
 _history = History()
 
 
+def auth_required(f):
+    @wraps(f)
+    def wrapper(*args, **kwargs):
+        auth = request.authorization
+        if not auth:
+            if current_user.is_authenticated:
+                return f(*args, **kwargs)
+            else:
+                abort(401)
+        user = UserDataAccess().get_user(user_id=auth.username)
+        retrieved_pass = UserDataAccess().login_user(auth.username)
+        if user is None or sha256_crypt.verify(auth.password, retrieved_pass):
+            abort(401)
+        login_user(user)
+        return f(*args, **kwargs)
+
+    return wrapper
+
+
 @api.route('/api/datasets/<int:dataset_id>/tables/<string:table_name>', methods=['GET'])
+@auth_required
 def get_table(dataset_id, table_name):
+    if (data_loader.has_access(current_user.username, dataset_id)) is False:
+        return abort(403)
     start = request.args.get('start')
     length = request.args.get('length')
     order_column = int(request.args.get('order[0][column]'))
@@ -26,7 +54,10 @@ def get_table(dataset_id, table_name):
 
 
 @api.route('/api/datasets/<int:dataset_id>/share', methods=['GET'])
+@auth_required
 def get_access_table(dataset_id):
+    if (data_loader.has_access(current_user.username, dataset_id)) is False:
+        return abort(403)
     start = request.args.get('start')
     length = request.args.get('length')
     access_table_columns = ['id_user', 'role']
@@ -43,7 +74,10 @@ def get_access_table(dataset_id):
 
 
 @api.route('/api/datasets/<int:dataset_id>/tables/<string:table_name>/history', methods=['GET'])
+@auth_required
 def get_history(dataset_id, table_name):
+    if (data_loader.has_access(current_user.username, dataset_id)) is False:
+        return abort(403)
     start = request.args.get('start')
     length = request.args.get('length')
     search = request.args.get('search[value]')
@@ -61,7 +95,10 @@ def get_history(dataset_id, table_name):
 
 
 @api.route('/api/datasets/<int:dataset_id>/tables/<string:table_name>/rows', methods=['POST'])
+@auth_required
 def add_row(dataset_id, table_name):
+    if (data_loader.has_access(current_user.username, dataset_id)) is False:
+        return abort(403)
     values = dict()
     columns = list()
     for key in request.args:
@@ -74,14 +111,20 @@ def add_row(dataset_id, table_name):
 
 
 @api.route('/api/datasets/<int:dataset_id>/tables/<string:table_name>/rows', methods=['DELETE'])
+@auth_required
 def delete_row(dataset_id, table_name):
+    if (data_loader.has_access(current_user.username, dataset_id)) is False:
+        return abort(403)
     row_ids = [key.split('-')[1] for key in request.args]
     data_loader.delete_row(dataset_id, table_name, row_ids)
     return jsonify({'success': True}), 200
 
 
 @api.route('/api/datasets/<int:dataset_id>/tables/<string:table_name>/columns', methods=['POST'])
+@auth_required
 def add_column(dataset_id, table_name):
+    if (data_loader.has_access(current_user.username, dataset_id)) is False:
+        return abort(403)
     column_name = request.args.get('col-name')
     column_type = request.args.get('col-type')
     data_loader.insert_column(dataset_id, table_name, column_name, column_type)
@@ -89,7 +132,10 @@ def add_column(dataset_id, table_name):
 
 
 @api.route('/api/datasets/<int:dataset_id>/tables/<string:table_name>/columns', methods=['PUT'])
+@auth_required
 def update_column(dataset_id, table_name):
+    if (data_loader.has_access(current_user.username, dataset_id)) is False:
+        return abort(403)
     column_name = request.args.get('col-name')
     column_type = request.args.get('col-type')
     data_loader.update_column_type(dataset_id, table_name, column_name, column_type)
@@ -97,14 +143,20 @@ def update_column(dataset_id, table_name):
 
 
 @api.route('/api/datasets/<int:dataset_id>/tables/<string:table_name>/columns', methods=['DELETE'])
+@auth_required
 def delete_column(dataset_id, table_name):
+    if (data_loader.has_access(current_user.username, dataset_id)) is False:
+        return abort(403)
     column_name = request.args.get('col-name')
     data_loader.delete_column(dataset_id, table_name, column_name)
     return jsonify({'success': True}), 200
 
 
 @api.route('/api/datasets/<int:dataset_id>/tables/<string:table_name>/date-time-transformations', methods=['PUT'])
+@auth_required
 def transform_date_or_time(dataset_id, table_name):
+    if (data_loader.has_access(current_user.username, dataset_id)) is False:
+        return abort(403)
     column_name = request.args.get('col-name')
     operation_name = request.args.get('operation-name')
     date_time_transformer.transform(dataset_id, table_name, column_name, operation_name)
@@ -112,8 +164,11 @@ def transform_date_or_time(dataset_id, table_name):
 
 
 @api.route('/api/datasets/update-dataset-metadata', methods=['PUT'])
+@auth_required
 def update_dataset_metadata():
     dataset_id = request.args.get('ds-id')
+    if (data_loader.has_access(current_user.username, dataset_id)) is False:
+        return abort(403)
     new_name = request.args.get('ds-name')
     new_desc = request.args.get('ds-desc')
     data_loader.update_dataset_metadata(dataset_id, new_name, new_desc)
@@ -121,7 +176,10 @@ def update_dataset_metadata():
 
 
 @api.route('/api/datasets/<int:dataset_id>/update-metadata', methods=['PUT'])
+@auth_required
 def update_table_metadata(dataset_id):
+    if (data_loader.has_access(current_user.username, dataset_id)) is False:
+        return abort(403)
     old_table_name = request.args.get('t-old-name')
     new_table_name = request.args.get('t-name')
     new_desc = request.args.get('t-desc')
@@ -130,7 +188,10 @@ def update_table_metadata(dataset_id):
 
 
 @api.route('/api/datasets/<int:dataset_id>/tables/<string:table_name>/impute-missing-data', methods=['PUT'])
+@auth_required
 def impute_missing_data(dataset_id, table_name):
+    if (data_loader.has_access(current_user.username, dataset_id)) is False:
+        return abort(403)
     column_name = request.args.get('col-name')
     function = request.args.get('function')
     data_transformer.impute_missing_data(dataset_id, table_name, column_name, function)
@@ -138,7 +199,10 @@ def impute_missing_data(dataset_id, table_name):
 
 
 @api.route('/api/datasets/<int:dataset_id>/tables/<string:table_name>/export', methods=['PUT'])
+@auth_required
 def export_table(dataset_id, table_name):
+    if (data_loader.has_access(current_user.username, dataset_id)) is False:
+        return abort(403)
     # Maybe later we might add other types, but for now this is hardcoded to export as CSV
     filename = table_name + ".csv"
     path = UPLOAD_FOLDER + "/" + filename
@@ -153,12 +217,16 @@ def export_table(dataset_id, table_name):
 
 
 @api.route('/api/download/<string:filename>', methods=['GET'])
+@auth_required
 def download_file(filename):
     return send_from_directory(UPLOAD_FOLDER, filename, as_attachment=True)
 
 
 @api.route('/api/datasets/<int:dataset_id>/tables/<string:table_name>/show-raw-data', methods=['GET'])
+@auth_required
 def show_raw_data(dataset_id, table_name):
+    if (data_loader.has_access(current_user.username, dataset_id)) is False:
+        return abort(403)
     start = request.args.get('start')
     length = request.args.get('length')
     order_column = int(request.args.get('order[0][column]'))
@@ -174,7 +242,10 @@ def show_raw_data(dataset_id, table_name):
 
 
 @api.route('/api/datasets/<int:dataset_id>/tables/<string:table_name>/find-and-replace', methods=['PUT'])
+@auth_required
 def find_and_replace(dataset_id, table_name):
+    if (data_loader.has_access(current_user.username, dataset_id)) is False:
+        return abort(403)
     colomn = request.args.get('col-name')
     replacement_function = request.args.get('replacement-function')
 
@@ -191,14 +262,20 @@ def find_and_replace(dataset_id, table_name):
 
 
 @api.route('/api/datasets/<int:dataset_id>/tables/<string:table_name>/normalize', methods=['PUT'])
+@auth_required
 def normalize(dataset_id, table_name):
+    if (data_loader.has_access(current_user.username, dataset_id)) is False:
+        return abort(403)
     column_name = request.args.get('col-name')
     numerical_transformer.normalize(dataset_id, table_name, column_name)
     return jsonify({'success': True}), 200
 
 
 @api.route('/api/datasets/<int:dataset_id>/tables/<string:table_name>/discretize', methods=['PUT'])
+@auth_required
 def discretize(dataset_id, table_name):
+    if (data_loader.has_access(current_user.username, dataset_id)) is False:
+        return abort(403)
     column_name = request.args.get('col-name')
     discretization = request.args.get('discretization')
     try:
@@ -221,7 +298,10 @@ def discretize(dataset_id, table_name):
 
 
 @api.route('/api/datasets/<int:dataset_id>/tables/<string:table_name>/outliers', methods=['PUT'])
+@auth_required
 def outliers(dataset_id, table_name):
+    if (data_loader.has_access(current_user.username, dataset_id)) is False:
+        return abort(403)
     try:
         column_name = request.args.get('col-name')
         option = request.args.get('option') == 'less-than'
@@ -234,7 +314,10 @@ def outliers(dataset_id, table_name):
 
 
 @api.route('/api/datasets/<int:dataset_id>/tables/<string:table_name>/rename-column', methods=['PUT'])
+@auth_required
 def rename_column(dataset_id, table_name):
+    if (data_loader.has_access(current_user.username, dataset_id)) is False:
+        return abort(403)
     to_rename = request.args.get('col-name')
     new_name = request.args.get('new-name')
     data_loader.rename_column(dataset_id, table_name, to_rename, new_name)
@@ -242,7 +325,10 @@ def rename_column(dataset_id, table_name):
 
 
 @api.route('/api/datasets/<int:dataset_id>/tables/<string:table_name>/chart', methods=['GET'])
+@auth_required
 def chart(dataset_id, table_name):
+    if (data_loader.has_access(current_user.username, dataset_id)) is False:
+        return abort(403)
     column_name = request.args.get('col-name')
     column_type = request.args.get('col-type')
 
@@ -253,9 +339,12 @@ def chart(dataset_id, table_name):
     else:
         return jsonify(numerical_transformer.chart_data_numerical(dataset_id, table_name, column_name))
 
-      
+
 @api.route('/api/datasets/<int:dataset_id>/tables/<string:table_name>/one-hot-encode-column', methods=['PUT'])
+@auth_required
 def one_hot_encode(dataset_id, table_name):
+    if (data_loader.has_access(current_user.username, dataset_id)) is False:
+        return abort(403)
     column_name = request.args.get('col-name')
     one_hot_encoder.encode(dataset_id, table_name, column_name)
     return jsonify({'success': True}), 200
